requirements.txt
resend==0.7.2
python-dotenv==1.0.0 